src/main/java/com/mentoring/mentoringprj/controller/TransactionController.java
package com.mentoring.mentoringprj.controller;

import com.mentoring.mentoringprj.domain.Transaction;
import com.mentoring.mentoringprj.exceptions.TransactionReadException;
import com.mentoring.mentoringprj.service.TransactionService;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
@RestController
@RequestMapping("/transactions")
public class TransactionController {

    private final TransactionService transactionService;

    public TransactionController(TransactionService transactionService) {
        this.transactionService = transactionService;
    }

    @GetMapping( path = "", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Transaction> getTransactions() throws TransactionReadException {
        return transactionService.getTransactions();
    }
}
